savetrend: use correct value for InstagramPublishedAt

+++ src/instagram/model_media.go
@@ -112,7 +112,7 @@ type MediaInfo struct {
 	OriginalWidth                int           `json:"original_width"`
 	PhotoOfYou                   bool          `json:"photo_of_you"`
 	Pk                           int64         `json:"pk"`
-	TakenAt                      float64       `json:"taken_at"`
+	TakenAt                      int64         `json:"taken_at"`
 	User                         struct {
 		FullName                   string `json:"full_name"`
 		HasAnonymousProfilePicture bool   `json:"has_anonymous_profile_picture"`
